% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_mdl.R
\name{best_mdl}
\alias{best_mdl}
\title{Selects model with minimal mean absolute forecating error.}
\usage{
best_mdl(...)
}
\arguments{
\item{...}{Arguments passed to \code{crossv_ts_multi}.}
}
\value{
Name of model function which produced smallest error.
}
\description{
Selects model with minimal mean absolute forecating error.
}
