% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossv_ts.R
\name{crossv_ts}
\alias{crossv_ts}
\title{Time series cross-validation}
\usage{
crossv_ts(y, forecastfunction, h = 1, tail2CV = NULL, ...)
}
\arguments{
\item{y}{Time series object.}

\item{forecastfunction}{Forecasting function. It must be able take ts object as first parameter and h (forecasting horizon) as second parameter.}

\item{h}{Forecast horizon.}

\item{tail2CV}{Integer. It specifies number of time series elements for those predicitions will be build.}

\item{...}{Additional parameters to forecastfunction.}
}
\value{
A time series object containing
  absolute prediction errors.
}
\description{
Computes the forecast errors obtained by applying forecastfunction to subsets
of the time series y using a rolling (accumulating) forecast origin. It
differs from forecast::tsCV by skipping head of the time series.
}
