% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossv_ts_multi.R
\name{crossv_ts_multi}
\alias{crossv_ts_multi}
\title{Time series cross-validation for list of modelling functions}
\usage{
crossv_ts_multi(ts, h, tail2CV, models, params)
}
\description{
Computes the forecast errors obtained by applying each of models to subsets
of the time series ts using a rolling (accumulating) forecast origin.
}
\examples{

crossv_ts_multi(gas, h = 1,
               tail2CV = 6,
               models = c("auto.arima", "ets", "bats"),
               params = list(
                 list(stepwise = FALSE,
                      approximation = FALSE),
                 list(damped = TRUE),
                 list()
               )
)
}
