% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ts_mdl.R
\name{fit_ts_mdl}
\alias{fit_ts_mdl}
\title{Fits forecast model based on time series, forecasting function and its
parameters.}
\usage{
fit_ts_mdl(ts = NULL, model = NULL, params = NULL)
}
\arguments{
\item{ts}{Time series object.}

\item{model}{Time series model function.}

\item{params}{List of parameters for the model function. If skipped than it
is taken from \code{frcst_fit_params} function.}
}
\value{
Time series model object \code{ets}, \code{auto.arima}, etc.
}
\description{
Fits forecast model based on time series, forecasting function and its
parameters.
}
