% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frcst_fit_params.R
\name{frcst_fit_params}
\alias{frcst_fit_params}
\title{Returns parameters for fitting of selected forecasting model}
\usage{
frcst_fit_params(model = NULL)
}
\arguments{
\item{model}{"auto.arima", "ets" or "bats".}
}
\value{
List with parameters.
}
\description{
Returns parameters for fitting of selected forecasting model
}
