% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_excel.R
\name{import_excel}
\alias{import_excel}
\title{Imports an Excel file with time series data.}
\usage{
import_excel(filepath = NULL)
}
\arguments{
\item{filepath}{Path to excel file}
}
\description{
Imports an Excel file with time series data.
}
