% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lwst_ts.R
\name{lwst_ts}
\alias{lwst_ts}
\title{Selects time series with lowest average value}
\usage{
lwst_ts(ts_lst)
}
\arguments{
\item{ts_lst}{Named list with time series.}
}
\value{
Name of the time series with lowest average absolute value.
}
\description{
Selects time series with lowest average value
}
