% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_column2tbl.R
\name{raw_column2tbl}
\alias{raw_column2tbl}
\title{Converts text column with numbers to a tibble}
\usage{
raw_column2tbl(raw_column = NULL, dec = NULL, startdate = NULL)
}
\description{
Converts text column with numbers to a tibble
}
