% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastbox.R
\name{forecastbox}
\alias{forecastbox}
\title{Функция запуска Shiny-приложения для загрузки и обработки Excel файла}
\usage{
forecastbox(port = getOption("shiny.port"), browser = TRUE,
  host = "0.0.0.0")
}
\arguments{
\item{port}{TCP порт, который слушает приложение. Если порт не задан, и нет опции shiny.port, то будет использован случайный порт.}

\item{browser}{Запускать ли браузер (логическая, TRUE по умолчанию)}

\item{host}{С каких компьютеров принимать запросы. По умолчанию со всех ("0.0.0.0"). Чтобы ограничить только локальной машиной, установите значение "127.0.0.1".}
}
\description{
Функция запуска Shiny-приложения для загрузки и обработки Excel файла
}
